requirements.txt
# Data Processing
pandas>=2.1.1
numpy>=1.26.0
lxml>=4.9.3
pdfplumber>=0.10.2
beautifulsoup4>=4.12.2
requests>=2.31.0

# Database
neo4j>=5.14.0

# Machine Learning & NLP
transformers>=4.34.0
torch>=2.1.0
scikit-learn>=1.3.1
sentence-transformers>=2.2.2

# Web UI
streamlit>=1.27.2
plotly>=5.17.0

# API
fastapi==0.103.2
uvicorn==0.23.2
pydantic>=2.7.4
python-dotenv>=0.21.0

# Testing
pytest>=7.4.2
pytest-cov>=4.1.0

# Development
black==23.9.1
isort==5.12.0
flake8==6.1.0
mypy==1.5.1
